+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Mark service `plan` as a required field
 - Add `minumim`, `maximum` validations for `number` type
+- Move helm charts to the operator repository
+- Add helm charts generator
 
 ## v0.9.0 - 2023-03-03